refactor resource
